# packages/contracts/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/characters/{character_id}/loadouts/{preset_id}": {
      "get": {
        "operationId": "EquipmentController_getPreset",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "preset_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadoutPresetEnvelope"
                }
              }
            }
          }
        },
        "summary": "读取装备预设",
        "tags": [
          "characters"
        ]
      },
      "put": {
        "operationId": "EquipmentController_savePreset",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "preset_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadoutPresetSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadoutPresetEnvelope"
                }
              }
            }
          }
        },
        "summary": "保存装备预设",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/v1/characters/{character_id}/loadouts/{preset_id}/equip": {
      "post": {
        "operationId": "EquipmentController_equipPreset",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "preset_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadoutPresetEnvelope"
                }
              }
            }
          }
        },
        "summary": "启用装备预设",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/v1/characters/{character_id}/queue": {
      "get": {
        "operationId": "QueueController_getQueue",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeQueue"
                }
              }
            }
          }
        },
        "summary": "读取角色闭关队列",
        "tags": [
          "queue"
        ]
      },
      "put": {
        "operationId": "QueueController_save",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueuePlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeQueueMutation"
                }
              }
            }
          }
        },
        "summary": "保存闭关队列",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/v1/characters/{character_id}/queue/preview": {
      "post": {
        "operationId": "QueueController_preview",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueuePlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeQueuePreview"
                }
              }
            }
          }
        },
        "summary": "预览闭关队列，不改变状态或资产",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/v1/characters/{character_id}/queue/pause": {
      "post": {
        "operationId": "QueueController_pause",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeQueueMutation"
                }
              }
            }
          }
        },
        "summary": "在周期边界暂停闭关队列",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/v1/characters/{character_id}/queue/resume": {
      "post": {
        "operationId": "QueueController_resume",
        "parameters": [
          {
            "name": "character_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeQueueMutation"
                }
              }
            }
          }
        },
        "summary": "恢复闭关队列",
        "tags": [
          "queue"
        ]
      }
    },
...
    "/api/v1/characters/{character_id}/dungeon-opportunities": {
      "get": {
        "tags": [
          "dungeon"
        ],
        "summary": "读取角色秘境机会",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色秘境机会信息。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonOpportunity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/dungeon-opportunities/teaching-grant": {
      "post": {
        "tags": [
          "dungeon"
        ],
        "summary": "领取教学赠送的秘境机会",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "教学赠送结果。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonOpportunity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/dungeon-runs": {
      "post": {
        "tags": [
          "dungeon"
        ],
        "summary": "创建秘境运行并消耗一次机会",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DungeonRunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "秘境运行已创建。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{character_id}/dungeon-runs/{run_id}": {
      "get": {
        "tags": [
          "dungeon"
        ],
        "summary": "读取秘境运行",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "秘境运行详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dungeons/{dungeon_id}/preview": {
      "post": {
        "tags": [
          "dungeon"
        ],
        "summary": "预览秘境运行",
        "parameters": [
          {
            "name": "dungeon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DungeonPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "秘境预览结果。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DungeonPreviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dungeon-runs/{run_id}": {
      "get": {
        "tags": [
          "dungeon"
        ],
        "summary": "读取秘境运行当前状态",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "秘境运行详情。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dungeon-runs/{run_id}/choices": {
      "post": {
        "tags": [
          "dungeon"
        ],
        "summary": "选择秘境路线",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DungeonChoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "秘境运行已更新。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dungeon-runs/{run_id}/finalize": {
      "post": {
        "tags": [
          "dungeon"
        ],
        "summary": "结算秘境运行",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "秘境运行已结算。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelopeDungeonRun"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "洞天 API",
    "description": "洞天服务端 REST API。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LoadoutPresetSaveRequest": {
        "type": "object",
        "required": [
          "expected_state_version",
          "name",
          "strategy_id"
        ],
        "properties": {
          "expected_state_version": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "weapon_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "armor_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "accessory_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "combat_consumables": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "item_id",
                "quantity"
              ],
              "properties": {
                "item_id": {
                  "type": "string"
                },
                "quantity": {
                  "oneOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "string"
                    }
                  ]
                }
              }
            }
          },
          "strategy_id": {
            "type": "string"
          }
        }
      },
      "LoadoutPreset": {
        "type": "object",
        "required": [
          "character_id",
          "preset_id",
          "name",
          "active",
          "complete",
          "state_version",
          "weapon_instance_id",
          "armor_instance_id",
          "accessory_instance_id",
          "combat_consumables",
          "strategy_id",
          "version"
        ],
        "properties": {
          "character_id": {
            "type": "string",
            "format": "uuid"
          },
          "preset_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "complete": {
            "type": "boolean"
          },
          "effective_next_cycle": {
            "type": "boolean"
          },
          "state_version": {
            "type": "integer"
          },
          "weapon_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "armor_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "accessory_instance_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "combat_consumables": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "item_id",
                "quantity"
              ],
              "properties": {
                "item_id": {
                  "type": "string"
                },
                "quantity": {
                  "type": "string"
                }
              }
            }
          },
          "strategy_id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "LoadoutPresetEnvelope": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LoadoutPreset"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "SuccessEnvelopeQueue": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Queue"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SuccessEnvelopeQueuePreview": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueuePreview"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SuccessEnvelopeQueueMutation": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueueMutation"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "QueuePlanRequest": {
        "type": "object",
        "required": [
          "expected_queue_version",
          "entries",
          "fallback"
        ],
        "properties": {
          "expected_queue_version": {
            "oneOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "string",
                "pattern": "^(?:0|[1-9]\\d*)$"
              }
            ]
          },
          "entries": {
            "type": "array",
            "maxItems": 3,
            "items": {
              "type": "object",
              "required": [
                "client_entry_id",
                "action_id",
                "mode",
                "on_blocked"
              ],
              "properties": {
                "client_entry_id": {
                  "type": "string"
                },
                "action_id": {
                  "type": "string"
                },
                "mode": {
                  "type": "string",
                  "enum": [
                    "COUNT",
                    "DURATION",
                    "UNTIL_INVENTORY",
                    "INFINITE"
                  ]
                },
                "target_value": {
                  "oneOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "string"
                    }
                  ]
                },
                "condition_item_id": {
                  "type": "string"
                },
                "condition_operator": {
                  "type": "string"
                },
                "on_blocked": {
                  "type": "string",
                  "enum": [
                    "SKIP",
                    "FALLBACK"
                  ]
                }
              }
            }
          },
          "fallback": {
            "type": "object",
            "required": [
              "action_id",
              "mode"
            ],
            "properties": {
              "action_id": {
                "type": "string"
              },
              "mode": {
                "type": "string",
                "enum": [
                  "INFINITE"
                ]
              }
            }
          }
        }
      },
      "QueueVersionRequest": {
        "type": "object",
        "required": [
          "expected_queue_version"
        ],
        "properties": {
          "expected_queue_version": {
            "oneOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "string",
                "pattern": "^(?:0|[1-9]\\d*)$"
              }
            ]
          }
        }
      },
...
      "DungeonRunCreateRequest": {
        "type": "object",
        "required": [
          "dungeon_id",
          "loadout_preset_id",
          "strategy_preset_id",
          "initial_route_id",
          "expected_state_version",
          "config_version"
        ],
        "properties": {
          "dungeon_id": {
            "type": "string"
          },
          "loadout_preset_id": {
            "type": "string"
          },
          "strategy_preset_id": {
            "type": "string"
          },
          "initial_route_id": {
            "type": "string"
          },
          "expected_state_version": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          },
          "config_version": {
            "type": "string"
          }
        }
      },
...
      "SuccessEnvelopeDungeonOpportunity": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DungeonOpportunityResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SuccessEnvelopeDungeonRun": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DungeonRunResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "DungeonPreviewRequest": {
        "type": "object",
        "required": [
          "character_id",
          "loadout_preset_id",
          "strategy_preset_id",
          "initial_route_id"
        ],
        "properties": {
          "character_id": {
            "type": "string"
          },
          "loadout_preset_id": {
            "type": "string"
          },
          "strategy_preset_id": {
            "type": "string"
          },
          "initial_route_id": {
            "type": "string"
          }
        }
      },
      "DungeonChoiceRequest": {
        "type": "object",
        "required": [
          "choice_id",
          "expected_run_version"
        ],
        "properties": {
          "choice_id": {
            "type": "string"
          },
          "expected_run_version": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          }
        }
      },
      "DungeonPreviewResponse": {
        "type": "object",
        "required": [
          "character",
          "dungeon",
          "config_version",
          "calculation_as_of"
        ],
        "properties": {
          "character": {
            "$ref": "#/components/schemas/DungeonOpportunityCharacter"
          },
          "dungeon": {
            "type": "object",
            "required": [
              "dungeon_id",
              "recommended_power",
              "base_success_rate",
              "estimated_success_rate",
              "choice_timeout_seconds",
              "opportunity_cost",
              "entry_items",
              "choices",
              "core_rewards"
            ],
            "properties": {
              "dungeon_id": {
                "type": "string"
              },
              "recommended_power": {
                "type": "string"
              },
              "base_success_rate": {
                "type": "string"
              },
              "estimated_success_rate": {
                "type": "string"
              },
              "choice_timeout_seconds": {
                "type": "integer"
              },
              "opportunity_cost": {
                "type": "integer"
              },
              "entry_items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              },
              "choices": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              },
              "core_rewards": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "config_version": {
            "type": "string"
          },
          "calculation_as_of": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}